hg: fix authormap name corruption in hgsubversion
Summary:
The AuthorMap in hgsubversion is currently derived from BaseMap which
has some strange behavior. During long running conversions, this caused some
names to be mangled.  For instance, if a long running conversion encounter a
commit with author 'abc' with converted name 'abc@uuid' and then later an author
'abcx' came along, it would convert the later name to 'abc@uuidx' because
BaseMap would match the prefix 'abc' with the beginning of the author name and
then reuse 'abc@uuid'. It then swapped the substring 'abc' from 'abcx' with
'abc@uuid' which resulted in the corrupt 'abc@uuidx'.

The fix is to add beginning and end markers to the regex name. The regex support
seems to be based around the ability to define custom regular expressions in
files to manipulate author/branch/tag names. In those cases the dictionary
__setitem__ call is passed a regular expression as a value. Since this patch
only affects the case where a string is passed, the regular expression scenario
will still work.

Reviewed By: singhsrb

Differential Revision: D7592383

fbshipit-source-id: 250b32d56782a6ec905cbd6620e0289fbdb523f0