server: gettreepack method preparation to return remapped Root Manifest hashes
Summary:
This should be mostly a no-op change, only the types change so that it prepares this code to be able to remap Root Manifest ids.
There is one small logical change that I am also shipping here, which is that fetching of linknode is happening in parallel with fetching of parents and content of the entry.

Reviewed By: StanislavGlebik

Differential Revision: D7600960

fbshipit-source-id: 2233f1c7d6f14e38cb075b5d7f40339d61cbf574