mononoke-types: change context keys to lowercase strings
Summary:
This is going to make life more convenient in upcoming diffs, when these strings are going to be used to generate blobstore keys.

The switch to lowercase is to make those blobstore keys look nicer.

The switch to string literals is because Rust really supports strings a lot better than it does bytestrings. In particular, `concat!` for literals works with strings, but there's no equivalent for bytestring literals. For current purposes the two are equivalent -- there's no runtime cost to using a string literal here.

Reviewed By: StanislavGlebik

Differential Revision: D7598616

fbshipit-source-id: f2fa543429921cca5fd8dc354ea11000fdab3f76