commitcloud: push commits when syncing
Summary:
Make commitcloud push unsynced commits to the server before modifying the cloud
state.

At the moment, this uses an infinitepush push, calling the underlying push
functions directly (since commitcloud knows exactly which commits to push,
there is no point performing discovery to calculate them).  Infinitepush pushes
require the full draft stack to be included, but in the future, this will be
replaced with an even smaller push of just the changed commits to Mononoke.

The re-use of infinitepush's internal functions is a bit hacky.  Later on I
will refactor the common functionality out to a utility library.

Reviewed By: StanislavGlebik

Differential Revision: D7535890

fbshipit-source-id: 975028a63bd6d9377a37e224a0cfd95ae1f1e16d