infinitepush: do not expose word 'infinitepush' to users and make it clear that all that messages are about autobackup (in background)
Summary:
Make it clear that all that messages are about autobackup (in background)

Even if it is disabled, manual `hg pushbackup` works

Also, since we announced publicly about renaming infinitepush we should not use it in user facing messages.

Also, that commands enable/disable background cloudsync, probably later we should check if cloudsync is enabled and improve messages to be more precise.

Reviewed By: StanislavGlebik

Differential Revision: D7635640

fbshipit-source-id: 2461b1b45e181c6e24b59f707a17f2fd60d4bb0e