improve smartlog for Commit Cloud user
Summary:
{F123775861}

Improve smartlog extension for Commit Cloud Users.

The difference with the current smartlog implementation is that in Commit Cloud world
a commit can be modified by a sequence of operations on another host and smartlog doesn't handle it.

After `hg cloudsync` run command we should be able to 'explain' in smartlog what has happened

this is important for user who opt out from 'updateonmove' feature because they will see two version of their stack all the time.

Reviewed By: DurhamG, quark-zju

Differential Revision: D7492969

fbshipit-source-id: 9ac2180f9abaa9ae596620b7f25d9ad8212deb28