pullbackup: fix pullbackups with --reporoot
Summary:
This was broken by D7604500 because the format for infinitepush
bookmark names doesn't expect a double slash when the reporoot begins with a
slash.  For instance, the bookmark name is
`infinitepush/backups/UNIXNAME/HOST/path/to/repo/heads/...` but we were querying
for `infinitepush/backups/UNIXNAME/HOST//path/to/repo/heads...`.

Reviewed By: singhsrb

Differential Revision: D7647924

fbshipit-source-id: 11abc38e44998d9164b474b9acce7bef7a011fe1