hg: fix smartlog.useancestorcache in shared workdirs
Summary:
The ancestorcache function requires a db in the cache directory, but if
the cache directory did not exist, the db would fail to get created, resulting
in slow smartlogs forever. Let's ensure the cache directory exists first.

This also switches to using cachevfs, which causes it to use the source repo's
cache directory instead of the shared one, which will mean better cache hits.

Reviewed By: quark-zju

Differential Revision: D7664296

fbshipit-source-id: b99acc22ba4f87af6e1edac1c9f15d39e8451ecc