asyncmemo: remove overgeneric Weight implementation
Summary:
That prevents us from implementing, for example, Weight for Bytes. There are no
impl specialization in Rust yet, and since Bytes do not implement HeapSizeOf,
we can't implement Weight for Bytes.

However we want to use Asyncmemo with Bytes. So in order to fix the issue let's
remove overgeneric Weight implementation and add implementation for the most
basic things.

Reviewed By: jsgf

Differential Revision: D7639162

fbshipit-source-id: b81d7e50599fdd8bdb8f9903d7aedff71734cdf2