ensure that we send no more than IOV_MAX iovecs at once
Summary:
Update the UnixSocket code to limit the number of iovecs that we send at one
time to folly::kIovMax.  (This constant is set from from IOV_MAX on platforms
that provide this setting.)

The thrift serializer code emits data into 16kB chunks, so if we had many
megabytes of data to send we could end up with thousands of iovecs.  The kernel
would fail the send with EMSGSIZE in this case.

Reviewed By: chadaustin

Differential Revision: D7665147

fbshipit-source-id: b7a60238d3fca973604b9037f22883cae80891ab