pullbackups: fix phases coming back as public
Summary:
There was a bug where hg pullbackups could result in commits being
marked as public when they shouldn't have. The issue was the the server side
loop iterates over heads, inspecting bundles as it goes, but it would only mark
the phase root for the head it was currently processing. While processing a
given head's bundle, it would take note of any other heads in the bundle and
mark them for bundling as well, but it wouldn't take note of their phase roots.

The fix is to just record all the heads found in that bundle and use them all in
the draft computation.

Reviewed By: ikostia

Differential Revision: D7648746

fbshipit-source-id: 93e6feb903fea49b6312c58dce4a6208105d002f