p4syncimport: read file contents from different sources
Summary:
Currently, the sync importer only reads files from perforce to ensure that the contents are up to date. Connecting to perforce and use `p4 print` is pretty slow, and we would like to use alternatives like `rcs` and `gzip` to speed things up, which are also what seqimporter is using.
This diff changes the reading behavior of sync importer so that it can also read from `rcs` and `gzip`.

A brief test on reading files from `Tools/...` shows that this works and the speedup is pretty good.

Differential Revision: D7619304

fbshipit-source-id: 563d0e40bcf7fa9a187cd3ede70878dccef4f9e9