add a basic restart command
Summary:
Add a pretty basic `restart` command to the Eden CLI.

This is relatively simple at the moment.  It has `--graceful` and `--full`
options, and defaults to a full restart at the moment if neither of these
arguments is specified.  We probably could add most of the arguments from the
`daemon` command to `restart`, but I'm not sure if we really want to support
that complexity in this command.

In the future we can make this command smarter and have it do a better job of
reporting the status as Eden restarts.

Reviewed By: chadaustin

Differential Revision: D7682838

fbshipit-source-id: 80c8d6604c3ca120fb71d6b8a388309218b5500c