infinitepush: don't push empty backups in new clones
Summary:
If the infinitepush backup state is empty (indicating this is probably a new
clone of a repository), and there are no commits or bookmarks to be backed up,
don't push a backupbundle that clears all the heads and bookmarks of any
backups that have the same servername and path.  This might be a reclone of the
repository, and the user might want to restore the backup.

When the first commit or bookmark is made in the new repository, the old
backup will be cleared at that point.

Differential Revision: D7669728

fbshipit-source-id: bec859bbbbee6b631efb4483a8fedbad8f3bcbac