mononoke: use exactly the same changesets during blobimport
Summary:
There are a few separate steps during blobimport. One of them is inserting the
blobs and another is inserting the changesets. That worked fine if RevlogRepo was
static. However if RevlogRepo has changed between first and second step we
could've inserted blobs for N commits, but insert (N + X) changesets into
changesets table. That means that a few commits would have no blobs at all.

This diff fixes it by fixing the changesets that we want to blobimport
beforehand.

Reviewed By: farnz

Differential Revision: D7615133

fbshipit-source-id: 1a66907e34a65588b101199c8f59abda53f7bc20