remove the `eden checkout` command
Summary:
This command directly invoked Eden's `checkOutRevision()` thrift call.  This
bypasses mercurial, so mercurial would not know about the working directory
parent change.

If we need to we could add this back as a debug command in the future.
However, this isn't something that normal users should ever invoke directly.

Reviewed By: chadaustin

Differential Revision: D7754529

fbshipit-source-id: 8d6aad70da37e56b3b8ced1ade4c8537141ab239