Change all assertions in GlobTest to go through one method.
Summary:
By introducing `_assert_glob()`, it eliminates a bit of the boilerplate in the
individual test cases. I think it also makes things easier to read because now
the glob patterns appear before the result of the glob.

Though most importantly, this will help with a subsequent change where I am
going to change the contract of the Thrift API, as now it can be done within
`_assert_glob()`.

Reviewed By: wez

Differential Revision: D7748871

fbshipit-source-id: 9609cde104979e892f0858e7a0c7e53e976ff8e2