have stop and full restart stop chg and buck
Summary:
These daemon processes tend to keep file handles open
and block shutdown in various annoying ways, so we terminate
them prior to stopping the eden server.

Note that the chg shutdown is linux specific.  We need a supported
command that does the right thing because this is fragile and won't
work on macos.  cc: quark-zju.

We'd talked about prompting the user about this; I can add prompting
as a follow on diff.

Reviewed By: chadaustin

Differential Revision: D7784903

fbshipit-source-id: 47a2f6395f99b2bfde3dbd049b6ca4804d011373