mononoke: fix delta application
Summary:
If baserev == Some(idx), we changed it to None. Also we can have None if
baserev == -1 in mercurial. However these two cases are different. In the first
case it means that we have a literal chunk, not delta. In the second case it
means that we have a delta against empty string! So this is technically almost
the same, except that delta against empty string also have a 12 bytes prefix.
Previously this prefix was used as part of a revision data.
This diff fixes it.

Reviewed By: quark-zju

Differential Revision: D7815713

fbshipit-source-id: def2e54b2cc7379ba8f931ecf3f3c0c38d716058