mononoke: make SshStream non-copyable and non-hashable
Summary:
Curently hgcli can connect to Mononoke only if they are on the same host,
because unix domain socket are used. Also Mononoke has to use separate unix
domain socket for different repos.
The goal of this stack of diffs is to remove these limitations:
1 Make it possible to have hgcli and Mononoke server on different hosts
2 Make it possible to use one port/unix domain socket to connect to many repos.

It's not required. And in the next diff we are going to add new SshStream
value, that will contain non-hashable and non-copyable struct in it.

Reviewed By: jsgf

Differential Revision: D7845160

fbshipit-source-id: 75bdd1706b1bd2f1b16a9e8c793c346a093bdd54