mononoke: small ssh_relay refactoring
Summary:
Curently hgcli can connect to Mononoke only if they are on the same host,
because unix domain socket are used. Also Mononoke has to use separate unix
domain socket for different repos.
The goal of this stack of diffs is to remove these limitations:
1 Make it possible to have hgcli and Mononoke server on different hosts
2 Make it possible to use one port/unix domain socket to connect to many repos.

This diff makes a small refactoring to make next diff smaller. There should be
no change in the functionality

Reviewed By: jsgf

Differential Revision: D7845155

fbshipit-source-id: 723a974e3e648b3e3200139aea27fc34691b8983