mononoke: check Preamble reponame
Summary:
Curently hgcli can connect to Mononoke only if they are on the same host, because unix domain socket are used. Also Mononoke has to use separate unix domain socket for different repos.
The goal of this stack of diffs is to remove these limitations:
1 Make it possible to have hgcli and Mononoke server on different hosts
2 Make it possible to use one port/unix domain socket to connect to many repos.

Check that that reponame from Preamble is the same as Mononoke reponame, and raise an error if it isn't.

Reviewed By: jsgf

Differential Revision: D7845157

fbshipit-source-id: 6567d5417ce58c4fba0364108592eed5bf0223c1